Move over to a new crate
