+++ mapbox/index.html
@@ -25,7 +25,7 @@
 	<h1>Demo</h1>
 	<div id="map"></div>
 	<script type="module">
-		import init, { PiggybackDemo } from "./pkg/widgetry_demo.js";
+		import init, { PiggybackDemo } from "./pkg/mapbox.js";
 
 		async function setup() {
 			// Initialize the WASM function.
