+++ widgetry_demo/Cargo.toml
@@ -9,7 +9,7 @@ crate-type = ["cdylib", "lib"]
 
 [features]
 default = ["widgetry/native-backend"]
-wasm = ["getrandom/js", "wasm-bindgen", "web-sys", "widgetry/wasm-backend"]
+wasm = ["getrandom/js", "wasm-bindgen", "widgetry/wasm-backend"]
 
 [dependencies]
 abstio = { path = "../abstio" }
@@ -21,5 +21,4 @@ rand = "0.8.3"
 rand_xorshift = "0.3.0"
 svg_face = "0.1.3"
 wasm-bindgen = { version = "0.2.70", optional = true }
-web-sys = { version = "0.3.47", optional = true }
 widgetry = { path = "../widgetry" }
