+++ widgetry_demo/src/lib.rs
@@ -530,7 +530,7 @@ fn make_controls(ctx: &mut EventCtx, tabs: &mut TabController) -> Panel {
                 .margin_above(30),
             Widget::row(vec![
                 Toggle::new_widget(
-                    true,
+                    false,
                     ctx.style()
                         .btn_outline
                         .text("Pause")
@@ -649,32 +649,4 @@ pub fn run_wasm(root_dom_id: String, assets_base_url: String, assets_are_gzipped
         .assets_base_url(assets_base_url)
         .assets_are_gzipped(assets_are_gzipped);
     run(settings);
-}
-
-// TODO Actually none of this has anything to do with widgetry_demo. Start a new crate to be clear.
-
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-pub struct PiggybackDemo {
-    render_only: widgetry::RenderOnly,
-    //draw_something: Drawable,
-}
-
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-impl PiggybackDemo {
-    pub fn create(gl: web_sys::WebGlRenderingContext) -> PiggybackDemo {
-        // Use this to initialize logging.
-        abstutil::CmdArgs::new().done();
-        let render_only = widgetry::RenderOnly::new(gl);
-        PiggybackDemo { render_only }
-    }
-
-    pub fn draw(&self) {
-        log::info!("Drawing...");
-        let mut g = self.render_only.gfx_ctx();
-        g.fork_screenspace();
-        g.draw_polygon(Color::RED, Polygon::rectangle(50.0, 50.0));
-        g.unfork();
-    }
 }
